+++ R/fit_lmms.R
@@ -22,8 +22,8 @@
 #' @param n.boots number of bootstrap samples to be used in the
 #' cluster bootstrap optimism correction procedure (CBOCP). If 0, no
 #' bootstrapping is performed
-#' @param n.cores number of cores to use to parallelize the computation
-#' of the CBOCP procedure. If \code{ncores = 1} (default), 
+#' @param n.cores number of cores to use to parallelize part of
+#' the computations. If \code{ncores = 1} (default), 
 #' no parallelization is done. Pro tip: you can use 
 #' \code{parallel::detectCores()} to check how many 
 #' cores are available on your computer
@@ -131,17 +131,7 @@ fit_lmms = function(y.names, fixefs, ranefs, long.data,
   if (n.cores > 1) {
     max.cores = parallel::detectCores()
     if (!is.na(max.cores)) {
-      diff = max.cores - n.cores
-      mess0 = paste('You requested', n.cores, 'cores for this computation.')
-      mess1 = paste(mess0, 'It seems that your computer actually has',
-                    max.cores, 'cores available.',
-                    'Consider increasing n.cores accordingly to speed computations up! =)')
-      if (diff > 0) warning(mess1, immediate. = TRUE)
-      mess2 = paste(mess0, 'However, seems that your computer only has',
-                    max.cores, 'cores available.',
-                    'Therefore, most likely computations will be performed using only', 
-                    max.cores, 'cores. =(')
-      if (diff < 0)  warning(mess2, immediate. = TRUE)
+      .check_ncores(avail = max.cores, requested = n.cores)
     }
   }
   
@@ -176,7 +166,8 @@ fit_lmms = function(y.names, fixefs, ranefs, long.data,
   # set up environment for parallel computing
   cl = parallel::makeCluster(n.cores)
   doParallel::registerDoParallel(cl)
- 
+  .info_ncores(n.cores, verbose = verbose)
+
   ########################
   ### original dataset ###
   ########################
@@ -247,9 +238,6 @@ fit_lmms = function(y.names, fixefs, ranefs, long.data,
   #######################
   if (n.boots >= 1) {
     if (verbose) cat('Bootstrap procedure started\n')
-    mess = ifelse(n.cores >=2, paste('in parallel, using', n.cores, 'cores'),
-                  'using a single core')
-    if (verbose) cat(paste('This computation will be run', mess, '\n'))
     # draw n bootstrap samples
     ids = unique(df$numeric.id)
     n = length(ids)