+++ R/pencox_baseline.R
@@ -16,7 +16,7 @@
 #' @param penalty the type of penalty function used for regularization.
 #' Default is \code{'ridge'}, other possible values are \code{'elasticnet'} 
 #' and \code{'lasso'}
-#' @param standardize logical argument: should the predicted random effects
+#' @param standardize logical argument: should the covariates
 #' be standardized when included in the penalized Cox model? Default is \code{TRUE}
 #' @param penalty.factor a single value, or a vector of values, indicating
 #' whether the covariates (if any) should be penalized (1) or not (0).
@@ -32,8 +32,7 @@
 #' in the bootstrap optimism correction procedure. If 0, no
 #' bootstrapping is performed
 #' @param n.cores number of cores to use to parallelize the computation
-#' of the bootstrap optimism correction procedure. If 
-#' \code{ncores = 1} (default), no parallelization is done. 
+#' of the CBOCP. If \code{ncores = 1} (default), no parallelization is done. 
 #' Pro tip: you can use \code{parallel::detectCores()} to check 
 #' how many cores are available on your computer
 #' @param verbose if \code{TRUE} (default and recommended value), information
@@ -152,17 +151,7 @@ pencox_baseline = function(data, formula,
   if (n.boots > 0) {
     max.cores = parallel::detectCores()
     if (!is.na(max.cores)) {
-      diff = max.cores - n.cores
-      mess0 = paste('You requested', n.cores, 'cores for this computation.')
-      mess1 = paste(mess0, 'It seems that your computer actually has',
-                    max.cores, 'cores available.',
-                    'Consider increasing n.cores accordingly to speed computations up! =)')
-      if (diff > 0) warning(mess1, immediate. = TRUE)
-      mess2 = paste(mess0, 'However, seems that your computer only has',
-                    max.cores, 'cores available.',
-                    'Therefore, most likely computations will be performed using only', 
-                    max.cores, 'cores. =(')
-      if (diff < 0)  warning(mess2, immediate. = TRUE)
+      .check_ncores(avail = max.cores, requested = n.cores)
     }
   }
   
@@ -234,12 +223,10 @@ pencox_baseline = function(data, formula,
   #######################
   if (do.bootstrap) {
     if (verbose) cat('Bootstrap procedure started\n')
-    mess = ifelse(n.cores >=2, paste('in parallel, using', n.cores, 'cores'),
-                                    'using a single core')
-    if (verbose) cat(paste('This computation will be run', mess, '\n'))
     # set up environment for parallel computing
     cl = parallel::makeCluster(n.cores)
     doParallel::registerDoParallel(cl)
+    .info_ncores(n.cores, verbose = verbose)
     
     # sample bootstrap row ids
     n = nrow(data)