+++ R/performance_pencox_baseline.R
@@ -11,8 +11,8 @@
 #' @param fitted_pencox the output of \code{\link{pencox_baseline}}
 #' @param times numeric vector with the time points at which
 #' to estimate the time-dependent AUC
-#' @param n.cores number of cores to use to parallelize the computation
-#' of the CBOCP procedure. If \code{ncores = 1} (default), 
+#' @param n.cores number of cores to use to parallelize part of
+#' the computations. If \code{ncores = 1} (default), 
 #' no parallelization is done. Pro tip: you can use 
 #' \code{parallel::detectCores()} to check how many 
 #' cores are available on your computer
@@ -137,19 +137,14 @@ performance_pencox_baseline = function(fitted_pencox, times = 1,
   if (n.boots > 0) {
     max.cores = parallel::detectCores()
     if (!is.na(max.cores)) {
-      diff = max.cores - n.cores
-      mess0 = paste('You requested', n.cores, 'cores for this computation.')
-      mess1 = paste(mess0, 'It seems that your computer actually has',
-                    max.cores, 'cores available.',
-                    'Consider increasing n.cores accordingly to speed computations up! =)')
-      if (diff > 0) warning(mess1, immediate. = TRUE)
-      mess2 = paste(mess0, 'However, seems that your computer only has',
-                    max.cores, 'cores available.',
-                    'Therefore, most likely computations will be performed using only', 
-                    max.cores, 'cores. =(')
-      if (diff < 0)  warning(mess2, immediate. = TRUE)
+      .check_ncores(avail = max.cores, requested = n.cores)
     }
   }
+  
+  # set up environment for parallel computing
+  cl = parallel::makeCluster(n.cores)
+  doParallel::registerDoParallel(cl)
+  .info_ncores(n.cores, verbose = verbose)
   
   #############################
   ##### NAIVE PERFORMANCE #####
@@ -172,7 +167,8 @@ performance_pencox_baseline = function(fitted_pencox, times = 1,
   # time-dependent AUC
   pmle.orig = as.numeric(coef(pcox.orig, s = 'lambda.min'))
   linpred.orig = X.orig %*% pmle.orig
-  tdauc.naive = foreach(i = 1:n.times, .combine = 'c') %do% {
+  tdauc.naive = foreach(i = 1:n.times, .combine = 'c',
+                        .packages = c('survivalROC')) %dopar% {
     auc = try(survivalROC(Stime = surv.data$time, 
                           status = surv.data$event, 
                           marker = linpred.orig, 
@@ -191,13 +187,7 @@ performance_pencox_baseline = function(fitted_pencox, times = 1,
   ###############################
   if (n.boots > 0) {
     if (verbose) cat('Computation of optimism correction started\n')
-    mess = ifelse(n.cores >=2, paste('in parallel, using', n.cores, 'cores'),
-                  'using a single core')
-    if (verbose) cat(paste('This computation will be run', mess, '\n'))
-    # set up environment for parallel computing
-    cl = parallel::makeCluster(n.cores)
-    doParallel::registerDoParallel(cl)
-    
+
     booty = foreach(b = 1:n.boots, .combine = 'rbind',
        .packages = c('survival', 'survcomp', 'survivalROC',
                      'glmnet', 'foreach')) %dopar% {
@@ -271,10 +261,7 @@ performance_pencox_baseline = function(fitted_pencox, times = 1,
       out$optimism = out$valid - out$train
       return(out)
     }
-    
-    # close the cluster
-    parallel::stopCluster(cl)
-    
+
     # compute the optimism correction for the C index
     c.vals = booty[booty$stat == 'C', ]
     c.opt = mean(c.vals$optimism, na.rm = TRUE)
@@ -295,6 +282,9 @@ performance_pencox_baseline = function(fitted_pencox, times = 1,
       cat('Computation of the optimism correction: finished :)\n')
     }
   }
+  
+  # close the cluster
+  parallel::stopCluster(cl)
 
   names(c.out) = c('n.boots', 'C.naive', 'cb.opt.corr', 'C.adjusted')
   names(tdauc.out) = c('pred.time', 'tdAUC.naive', 'cb.opt.corr', 'tdAUC.adjusted')